third_party/blink/web_tests/platform/generic/fast/dom/HTMLImageElement/image-srcset-in-template-expected.txt

PASS document.getElementById("canary").clientWidth is 75
PASS currentSrcFileName("canary") is "blue_rect.jpg?0"
PASS document.querySelector("#clone img").currentSrc.indexOf("image-set-1x") is not -1
PASS document.querySelector("#clone img").currentSrc.indexOf("blue_rect") is -1
PASS document.querySelector("#import img").currentSrc.indexOf("image-set-1x") is not -1
PASS document.querySelector("#import img").currentSrc.indexOf("blue_rect") is -1
PASS successfullyParsed is true

TEST COMPLETE
